requirements.txt
Flask==1.1.2
mysql-connector-python==8.0.23
flask-sqlalchemy==2.5.1
flask-login==0.5.0
pgeocode==0.3.0
coverage==6.3.2
pytest==7.1.0
Werkzeug~=1.0.1